feat: add commands.rs module, move activate task println to  command, closes #21
